Add files via upload
Code Task


The datasets consists of several medical predictor variables and one target variable, Outcome. 
Predictor variables includes the number of pregnancies the patient has had, their BMI, insulin level, age, and so on.

perform classification and build a machine learning model by performing standard scaling to accurately predict whether or not the patients in the 
dataset have diabetes or not?